Continue with the restructure
