+++ tidy3d/components/spice/__init__.py
@@ -1,0 +1,5 @@
+"""
+This directory includes shared component SPICE definitions that could be compiled from a RF lumped element or a TCAD component.
+The idea is that, ultimately, each component model can be compiled into an electrical element cleanly and used to create
+a SPICE model.
+"""
