--- tidy3d/components/tcad/mobility/__init__.py
@@ -1,3 +1,0 @@
-"""
-This directory corresponds to mobility models specifically.
-"""
