+++ tidy3d/components/tcad/source/generic.py
@@ -1,0 +1,3 @@
+
+
+HeatChargeSourceType = Union[HeatSource, HeatFromElectricSource, UniformHeatSource]
